PDFBOX-2576: remove unneeded unboxing
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1666299 13f79535-47bb-0310-9956-ffa450edef68

+++ xmpbox/src/main/java/org/apache/xmpbox/schema/PDFAIdentificationSchema.java
@@ -105,7 +105,7 @@ public class PDFAIdentificationSchema extends XMPSchema
      */
     public void setPart(Integer value)
     {
-        setPartValueWithInt(value.intValue());
+        setPartValueWithInt(value);
     }
 
     /**
